ver 3.0.0 stable - trying to re-organize and optimize
