+++ includes/functions.h
@@ -27,4 +27,4 @@ void login(void);
 
 #endif
 
-//ver 2.1.0 stable
+//ver 3.0.0 stable
