+++ src/functions.c
@@ -5,99 +5,8 @@
 #include "flats.h"
 #include "functions.h"
 
-flat flats;
+extern flat flats;
 
-//Creates memory space for owner field and returns a memory location
-Owner* create_owner(){
-    Owner* x = (Owner *)malloc(sizeof(Owner));
-    memset(x, 0, sizeof(char) * 100); //initializing the structure to 0 just incase, anyhow we'll be overwriting it in later codes
-    if(x == NULL){
-        printf("Memory allocation failed");
-        exit(1);
-    }
-    return x;
-}
-
-//Reads the file (data.csv) and saves it to the memory (residence array)
-void read_file() {
-    FILE *file = fopen(dataset, "r");
-    if (file == NULL) {
-        printf("error: couldn't open the file\n");
-        return;
-    }
-    char header[150];
-    fgets(header, sizeof(header), file);  // Ignorin the first line in the CSV file ('Cause it has headings, which is unnecessery)
-    char line[500];
-    char status[15];
-
-    while (fgets(line, sizeof(line), file)) { 
-        int i=0;
-        while(line[i]!=','){
-            status[i]= line[i];
-            i++;
-        }
-        status[i] = '\0';
-
-        if(strcmp(status, "Booked")==0){
-            flats.owner = create_owner(); 
-
-            if (sscanf(line, "%9[^,],%9[^,],%9[^,],%d,%49[^,],%49[^,],%f,%f,%f,%s", flats.status, flats.ID, flats.type, &flats.price, flats.owner->name, flats.owner->o_info, &flats.owner->paid, &flats.owner->bal, &flats.owner->due, flats.owner->date) == 10) {
-                if (current >= size) {
-                    printf("Error: Exceeded maximum residence capacity\n");
-                    break;
-                    }
-            residence[current++] = flats; // increasing the current value (from 0 (in dsa.c file)) till the loop ends... (It'll store the total number of Lines in the file and, residence array will move to the next structure element to save the data)
-            }
-        }
-        
-        else if (strcmp(status, "Available")==0){
-            if (sscanf(line, "%9[^,],%9[^,],%9[^,],%d", flats.status, flats.ID, flats.type, &flats.price) == 4) {
-                if (current >= size) {
-                    printf("Error: Exceeded maximum residence capacity\n");
-                    break;
-                    }
-            flats.owner = NULL;
-            residence[current++] = flats;
-            }
-        }
-        
-        else {
-            printf("Failed to parse line: %s\n", line);
-            printf("line: %s\n", line); //upon error reading the file, it'll just print this
-        }
-    }
-    fclose(file);
-    printf("\nLoading the main menu...\n");
-}
-//This one will overwrite the whole thing (i mean the file) - ie., we're using it in the end of the program to save the data that we fetched/edited during the runtime... so, we'll be erasing the previous data in .csv file and rewrite everything that we've inside the memory (residence array)
-void write_file(){
-    FILE *file = fopen(dataset, "w");
-        if (file == NULL){
-            printf("error: couldn't open the file");
-            return;
-        }
-    fprintf(file, "Status,Flat ID,Type,Price,Owner Name,Contact Info,Amount Paid So Far,Remaining Balance,Next Monthly Installment,Due Date\n");
-    char line[500];
-    for(int i=0; i<current; i++){
-        if(strcmp(residence[i].status, "Booked")==0){
-            sprintf(line, "%s,%s,%s,%d,%s,%s,%.2f,%.2f,%.2f,%s\n",residence[i].status, residence[i].ID, residence[i].type, residence[i].price, residence[i].owner->name, residence[i].owner->o_info, residence[i].owner->paid, residence[i].owner->bal, residence[i].owner->due, residence[i].owner->date);
-
-
-            fwrite(line, sizeof(char), strlen(line), file);
-            free(residence[i].owner);
-        }
-        else if(strcmp(residence[i].status, "Available")==0){
-            sprintf(line, "%s,%s,%s,%d,%p\n",residence[i].status, residence[i].ID, residence[i].type, residence[i].price, residence[i].owner);
-
-            fwrite(line, sizeof(char), strlen(line), file);
-        }
-        else{
-            printf("\nwriting error at %d\n", i);
-        }
-    }
-    fclose(file);
-
-}
 
 //calculates the downpayment amount by fetching price of the flat and the % of downpayment
 float down(int price, float value){
@@ -153,7 +62,7 @@ int book(char *id, char *name, char *contact){
     int ch;
     for(int i=0; i<current; i++){
         if(strcmp(residence[i].ID, id)==0){
-            residence[i].owner = create_owner(); //allocating memory for a new owner
+            residence[i].owner = createOwner(); //allocating memory for a new owner
             strcpy(residence[i].owner->name , name);
             strcpy(residence[i].owner->o_info , contact);
             printf("\nPayment method: 1. Full on payment    2. Pay with EMI\nEnter your choice: ");
@@ -166,7 +75,7 @@ int book(char *id, char *name, char *contact){
                     printf("\nSuccessfully created a record, in the name of %s with flat id %s\n", residence[i].owner->name, residence[i].ID);
                     strcpy(residence[i].status , "Booked");
                     residence[i].owner->paid = residence[i].price;
-                    residence[i].owner->bal = 00.00;
+                    residence[i].owner->balance = 00.00;
                     residence[i].owner->due = 00.00;
                     strcpy(residence[i].owner->date , "NULL");
                     return 1;
@@ -178,7 +87,7 @@ int book(char *id, char *name, char *contact){
                         strcpy(residence[i].owner->name , name);
                         strcpy(residence[i].owner->o_info , contact);
                         residence[i].owner->paid = down(residence[i].price, 20);
-                        residence[i].owner->bal = residence[i].price - residence[i].owner->paid;
+                        residence[i].owner->balance = residence[i].price - residence[i].owner->paid;
                         residence[i].owner->due = emi(principal(residence[i].price, down(residence[i].price, 20)), 7, 5);
                         strcpy(residence[i].owner->date , "1/1/2024");
                         printf("\nSuccessfully created a record, in the name of %s with flat id %s, your next due, %.2f, is on %s\n", residence[i].owner->name, residence[i].ID,residence[i].owner->due,residence[i].owner->date);
@@ -189,7 +98,7 @@ int book(char *id, char *name, char *contact){
                         strcpy(residence[i].owner->name , name);
                         strcpy(residence[i].owner->o_info , contact);
                         residence[i].owner->paid = down(residence[i].price, 25);
-                        residence[i].owner->bal = residence[i].price - residence[i].owner->paid;
+                        residence[i].owner->balance = residence[i].price - residence[i].owner->paid;
                         residence[i].owner->due = emi(principal(residence[i].price, down(residence[i].price, 25)), 7.5, 6);
                         strcpy(residence[i].owner->date , "1/1/2024");
                         printf("\nSuccessfully created a record, in the name of %s with flat id %s, your next due, %.2f, is on %s\n", residence[i].owner->name, residence[i].ID,residence[i].owner->due,residence[i].owner->date);
@@ -200,7 +109,7 @@ int book(char *id, char *name, char *contact){
                         strcpy(residence[i].owner->name , name);
                         strcpy(residence[i].owner->o_info , contact);
                         residence[i].owner->paid = down(residence[i].price, 30);
-                        residence[i].owner->bal = residence[i].price - residence[i].owner->paid;
+                        residence[i].owner->balance = residence[i].price - residence[i].owner->paid;
                         residence[i].owner->due = emi(principal(residence[i].price, down(residence[i].price, 30)), 8, 7);
                         strcpy(residence[i].owner->date , "1/1/2024");
                         printf("\n Successfully created a record, in the name of %s with flat id %s, your next due, %.2f, is on %s\n", residence[i].owner->name, residence[i].ID,residence[i].owner->due,residence[i].owner->date);
@@ -244,7 +153,7 @@ int flatinfo(char *id){
             printf("\n%-10s %-20s %-20s\n", " ", "Flat no.:", residence[i].ID);
             
             if(strcmp(residence[i].status, "Booked")==0){
-                if(residence[i].owner->bal == 0){
+                if(residence[i].owner->balance == 0){
                     printf("%-10s %-20s %-20s\n", " ", "Owner:", residence[i].owner->name);
                     printf("%-10s %-20s %-20s\n", " ", "Contact info:", residence[i].owner->o_info);
                     printf("%-10s %-20s %-20s\n", " ", "Type:", residence[i].type);
@@ -257,7 +166,7 @@ int flatinfo(char *id){
                     printf("%-10s %-20s %-20s\n", " ", "Contact info:", residence[i].owner->o_info);
                     printf("%-10s %-20s %-20s\n", " ", "Type:", residence[i].type);
                     printf("%-10s %-20s %-20d\n", " ", "Price:", residence[i].price);
-                    printf("%-10s %-20s %-20.f\n", " ", "Balance:", residence[i].owner->bal);
+                    printf("%-10s %-20s %-20.f\n", " ", "Balance:", residence[i].owner->balance);
                     printf("%-10s %-20s %-20.2f\n", " ", "Due Next Month:", residence[i].owner->due);
                     printf("%-10s %-20s %-20s\n", " ", "Due Date:", residence[i].owner->date);
                     return 2;
@@ -279,11 +188,11 @@ void pay(char *id){
     for(int i=0; i<current; i++){
         if(strcmp(id, residence[i].ID)==0){
             residence[i].owner->paid += residence[i].owner->due;
-            residence[i].owner->bal = residence[i].price - residence[i].owner->paid;
+            residence[i].owner->balance = residence[i].price - residence[i].owner->paid;
             residence[i].owner->due = 00.00;
             strcpy(residence[i].owner->date , "1/2/2024");
             printf("\nChanges recorded on Flat number %s\n\nUpdated details:\n", residence[i].ID );
-            printf("\n%-10s %-20s %-20.2f", " ", "Balance:", residence[i].owner->bal);
+            printf("\n%-10s %-20s %-20.2f", " ", "Balance:", residence[i].owner->balance);
             printf("\n%-10s %-20s %-20.2f", " ", "Due Next Month:", residence[i].owner->due);
             printf("\n%-10s %-20s %-20s\n", " ", "Due Date:", residence[i].owner->date);
             
@@ -592,4 +501,4 @@ void login(){
     }
 }
 
-//ver 2.1.0 stable
+//ver 3.0.0 stable
